포탑소환확률(완), 포탑강화(미완)
포탑 소환 확률 구현 완료 (확률 분포는 추후 수정) 포탑강화 시스템은 구현중

+++ Assets/Scripts/Tower/Projectile.cs
@@ -7,12 +7,12 @@ public class Projectile : MonoBehaviour
 {
     private Movement2D movement2D;
     private Transform target;
-    private int damage;
+    private float damage;
     private bool isSplashDamage;
     private int splashDamage;
     private float splashRadius;
 
-    public void Setup(Transform target,int damage, bool isSplashDamage = false, int splashDamage = 0, float splashRadius = 0f)
+    public void Setup(Transform target,float damage, bool isSplashDamage = false, int splashDamage = 0, float splashRadius = 0f)
     {
         movement2D=GetComponent<Movement2D>();
         this.target = target;  //타워가 설정해준 target
